Issue #290: move test_util into test
